Rename the config module and document
