+++ examples/siteConfig/QtSiteConfig.py
